curl: reverting to 7.54.1, as the new version causes rc to crash at boot time, causing a kernel panic.
rc/dsl_fb.o is linked against libcurl, and this is another closed source piece over which I have no
control...

Revert "curl: Updated to 7.55.1 (closes #1450)"

This reverts commit b8210c6e8e2aa442e285b3ceeb0d55a87761d8ac.

+++ release/src/router/curl/tests/server/sockfilt.c
@@ -549,7 +549,7 @@ static DWORD WINAPI select_ws_wait_thread(LPVOID lpParameter)
     free(data);
   }
   else
-    return (DWORD)-1;
+    return -1;
 
   /* retrieve the type of file to wait on */
   type = GetFileType(handle);
